@@ -7,17 +7,18 @@
     "build": "parcel build public/index.html"
   },
   "dependencies": {
-    "@fluent/bundle": "^0.17.0",
-    "@fluent/langneg": "^0.6.0",
-    "@fluent/sequence": "^0.7.0",
+    "@fluent/bundle": "^0.18.0",
+    "@fluent/langneg": "^0.7.0",
+    "@fluent/sequence": "^0.8.0",
     "cached-iterable": "^0.3.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
-    "parcel": "^2.5.0"
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "parcel": "^2.5.0",
+    "process": "^0.11.10"
   },
   "engines": {
     "node": ">=14.0.0",
