@@ -1,12 +1,12 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import { AppLocalizationProvider } from "./l10n";
 import { App } from "./App";
 
-ReactDOM.render(
+const root = createRoot(document.getElementById("root")!);
+root.render(
   <AppLocalizationProvider>
     <App />
-  </AppLocalizationProvider>,
-  document.getElementById("root")
+  </AppLocalizationProvider>
 );
